requirements.txt
astropy==5.0.4
astroquery==0.4.6
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
cryptography==37.0.2
distlib==0.3.4
filelock==3.7.0
html5lib==1.1
identify==2.5.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
jeepney==0.8.0
keyring==23.5.0
nodeenv==1.6.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
py==1.11.0
pycparser==2.21
pyerfa==2.0.0.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pyvo==1.3
PyYAML==6.0
requests==2.27.1
SecretStorage==3.3.2
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
validators==0.20.0
virtualenv==20.14.1
webencodings==0.5.1
zipp==3.8.0
ztfquery==1.18.2
